Testing Python open cv using image from google.
I took an image of a frisbee throw from google and tried to isolate just the hue of the frisbee so that it would be clearly visible. I couldn't get only the frisbee to show up without a lot of the surroundings also and when I tried to change the values of which hues to isolate either the same thing happened or nothing showed up. I think it's because the frisbee was white so I'm going to try to get my own image with a more colorful frisbee.