Frisbee using reflective stickers
I used reflective stickers I got off of amazon and put them in different places on a frisbee to test how easy it was to find them in a photo. This is the frisbee I will use to test outside. I put one sticker in the middle to hopefully be able to track rotation. I put three different colored sticker in roughly even positions around the frisbee to hopefully see which one shows up best and as another way to possivbe get rotation eventually.