Testing frisbee with reflective stickers
I tested the frisbee I made using reflective stickers outside. The frisbee doesn't have the stickers on the side I showed in the other picture because when the frisbee was thrown they wouldn't stay on. The sticker shows up some in the color image but it is too blurry in the black and white image to get any data out of. I thing the sticker would need to be bigger and more distinct to have a chance at getting meaningful data.